logger.go: Re-implement logging using a Go channel
Here we re-implement logging such that we construct a
new logger object that embeds a Go channel. The embedded
Go channel consumes typed logged messages and a coroutine
consumes them out the other end of the channel and writes
them out to the io.Writer the logger was constructed with.

V.2: rebased on top of master.

Signed-off-by: Edward O'Callaghan <quasisec@google.com>