Fix nil interface != nil issue (#37)
* Fix nil interface != nil issue

When `Init()` calls `setup()` and `setup()` fails, it would return a `nil` value of the type `*syslog.Writer`. Storing this inside an `io.Writer` interface type variable `il`, `wl`, or `el` inside `Init()` means that those variables will not be `nil` (because they are a valid `nil` value of the `io.Writer` interface). Thus the checks `if il != nil` etc. will always be true and the `nil` value `syslog.Writer` pointer will always be added to the `iLogs`, `wLogs`, `eLogs` lists. Because we're in a case where `setup()` returned an error, we'll attempt to log the error later on using the `io.MultiWriter(eLogs...)` which contains one `nil` value writer. This means the `syslog.(*Writer).Write()` call will fail with segmentation violation and lead to the program crashing.

With this commit, we change the interface of `setup()` to return `io.Writer` interface values directly which can be safely compared against `nil` and  thus the `nil` value writers are never added to the `io.MultiWriter`s.